News 34

+++ Modules/Spd/Category/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace Spd\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Spd\Category\Http\Requests\CategoryRequest;
+use Spd\Category\Models\Category;
 use Spd\Category\Repositories\CategoryRepo;
 use Spd\Category\Services\CategoryService;
 
@@ -20,44 +21,56 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize('index', Category::class);
         $categories = $this->repo->index()->paginate(10);
+
         return view('Category::index', compact('categories'));
     }
 
     public function create()
     {
+        $this->authorize('index', Category::class);
         $categories = $this->repo->index()->get();
+
         return view('Category::create', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('index', Category::class);
         $this->service->store($request);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ساخته شد']);
     }
 
     public function edit($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $categories = $this->repo->index()->where('id', '!=', $category->id)->get();
-
         return view('Category::edit', compact(['category', 'categories']));
+
     }
 
     public function update(CategoryRequest $request, $id)
     {
+        $this->authorize('index', Category::class);
         $this->service->update($request, $id);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ابدیت شد']);
     }
 
     public function destroy($id)
     {
+        $this->authorize('index', Category::class);
         $this->repo->delete($id);
+
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت حذف شد']);
     }
 
     public function changeStatus($id)
     {
+        $this->authorize('index', Category::class);
         $category = $this->repo->findById($id);
         $this->repo->changeStatus($category);
