+++ Modules/Spd/User/Http/Controllers/UserController.php
@@ -1,13 +1,14 @@
 <?php
 namespace Spd\User\Http\Controllers;
 
-use App\Http\Controllers\Controller;
-use Spd\Role\Repositories\RoleRepo;
-use Spd\User\Http\Requests\AddRoleRequest;
-use Spd\User\Http\Requests\UserRequest;
-use Spd\User\Http\Requests\UserUpdateRequest;
-use Spd\User\Repositories\UserRepo;
+use Spd\User\Models\User;
 use Spd\User\Services\UserService;
+use Spd\Role\Repositories\RoleRepo;
+use Spd\User\Repositories\UserRepo;
+use App\Http\Controllers\Controller;
+use Spd\User\Http\Requests\UserRequest;
+use Spd\User\Http\Requests\AddRoleRequest;
+use Spd\User\Http\Requests\UserUpdateRequest;
 
 
 
@@ -26,44 +27,65 @@ class UserController extends Controller
         $this->repo = $userRepo;
         $this->service = $userService;
     }
+
     public function index()
     {
+        $this->authorize('index', User::class);
         $users = $this->repo->index();
+
         return view('User::index', compact('users'));
     }
+
     public function create()
     {
+        $this->authorize('index', User::class);
         return view('User::create');
     }
+
     public function store(UserRequest $request)
     {
+        $this->authorize('index', User::class);
         $this->service->store($request);
+
         return to_route('users.index');
     }
+
     public function edit($id)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($id);
+
         return view('User::edit', compact('user'));
     }
+
     public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('index', User::class);
         $this->service->update($request, $id);
+
         return to_route('users.index');
     }
+
     public function destroy($id)
     {
+        $this->authorize('index', User::class);
         $this->repo->delete($id);
 
         return to_route('users.index')->with(['success_delete' => 'کاربر با موفقیت حذف شد']);
     }
+
+    // Role
     public function addRole($user_id, RoleRepo $roleRepo)
     {
+        $this->authorize('index', User::class);
         $roles = $roleRepo->index()->get();
+
         return view('User::add-roles', compact(['user_id', 'roles']));
     }
 
     public function addRoleStore(AddRoleRequest $request, $userId)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($userId);
         $this->service->addRole($request->role, $user);
 
@@ -73,10 +95,10 @@ class UserController extends Controller
 
     public function removeRole($userId, $roleId, RoleRepo $roleRepo)
     {
+        $this->authorize('index', User::class);
         $user = $this->repo->findById($userId);
         $role = $roleRepo->findById($roleId);
         $this->service->deleteRole($user, $role->name);
-
         alert()->success('حذف کردن مقام', 'عملیات با موفقیت انجام شد');
         return to_route('users.index');
     }
